Allow functions as shorthand commands (#118)
The documentation declares `command` as being of type string or Function, and states that it can be "omitted by directly setting the target with the command". It made me believe that I could directly pass a function as a target, but this actually only works for strings. This PR allows passing a function as target and omitting `command`.